% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_pairs}
\alias{make_pairs}
\title{Make pairs of PCs to plot}
\usage{
make_pairs(ind, how = c("step1", "step2", "all"))
}
\arguments{
\item{ind}{Indices of PCs to plot.}

\item{how}{Which pairs to make? For \code{1:4}, "step1" makes 1-2, 2-3, 3-4;
"step2" makes 1-2, 3-4, and "all" makes 1-2, 2-3, 3-4, 1-4, 2-4, 3-4.}
}
\value{
A two-column matrix of pairs of indices to plot.
}
\description{
Make pairs of PCs to plot
}
\examples{
make_pairs(1:4)
make_pairs(1:4, how = "step2")
make_pairs(1:5, how = "step2")
make_pairs(1:4, how = "all")
}
