% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{pc_deconv}
\alias{pc_deconv}
\title{Alternating deconvolution with incremental building.}
\usage{
pc_deconv(
  PC,
  m_exponent,
  use_varimax = TRUE,
  ind_plot = integer(0L),
  ncores = NULL,
  ...
)
}
\arguments{
\item{PC}{Matrix of PC scores, N x K.}

\item{m_exponent}{Exponent to be applied to the coefficients of \code{Q}.
A value between 3 and 30 is usually appropriate to priotize larger Q values.}

\item{use_varimax}{Whether to transform PCs using a varimax rotation?
Default is \code{TRUE}. It generally helps with the warm starts (initializations).}

\item{ncores}{Number of cores to use. Default is \code{1}. You can use e.g.
\code{bigparallelr::nb_cores()}.}

\item{...}{
  Arguments passed on to \code{\link[=pc_deconv_withstart]{pc_deconv_withstart}}
  \describe{
    \item{\code{conv_diff}}{Maximum difference between consecutive values of mixture
coefficients to stop updating these weights. Default is \code{1e-4}.}
    \item{\code{max_iter}}{Maximum number of iterations. Default is \code{200}.}
    \item{\code{thr_coef}}{Threshold under which coefficients of \code{Q} are not used.
Default is 0.6.}
  }}
}
\value{
A list of several \code{W}, after convergence of the deconvolution
\eqn{PC \approx Q \cdot W^T \cdot PC}, for all subsets of \code{PC} using from
\code{2} to \code{ncol(PC)} columns.
}
\description{
Alternating deconvolution with incremental building.
}
\examples{
PC <- prcomp(iris[1:4])$x
PC_ref_cheat <- do.call("rbind", by(PC, iris$Species, colMeans))  # cheating
all_PC_ref_conv <- pc_deconv(PC, m_exponent = 10)
plot(PC, pch = 20, cex = 1)
points(PC_ref_cheat, col = "orange", pch = 3, lwd = 2)
points(pc_refs(PC, all_PC_ref_conv[[3]]), col = "red", pch = 4, lwd = 2)
points(pc_refs(PC, all_PC_ref_conv[[4]]), col = "green", pch = 5, lwd = 2)
}
\seealso{
\link{pc_deconv_withstart}
}
