% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{pc_deconv_withstart}
\alias{pc_deconv_withstart}
\title{Alternating deconvolution with starting value.}
\usage{
pc_deconv_withstart(
  PC,
  PC_ref_init,
  m_exponent,
  thr_coef = 0.6,
  conv_diff = 1e-04,
  max_iter = 200,
  ...
)
}
\arguments{
\item{PC_ref_init}{Matrix of reference positions in the PCs (L x K, where L
is the number of references) to be used as \emph{starting values}.}

\item{m_exponent}{Exponent to be applied to the coefficients of \code{Q}.
A value between 3 and 30 is usually appropriate to priotize larger Q values.}

\item{thr_coef}{Threshold under which coefficients of \code{Q} are not used.
Default is \code{0.6}.}

\item{conv_diff}{Maximum difference between consecutive values of mixture
coefficients to stop updating these weights. Default is \code{1e-4}.}

\item{max_iter}{Maximum number of iterations. Default is \code{200}.}

\item{...}{
  Arguments passed on to \code{\link[=pc_mixtures]{pc_mixtures}}
  \describe{
    \item{\code{min_coef}}{Minimum non-zero mixture coefficient. Default is \code{1e-5}.
This is used in the first iteration of the algorithm only, so it can happen
that some coefficients in the results are lower (but this should be rare).}
    \item{\code{max_coef}}{Maximum mixture coefficient when only one coefficient is used.
Default is \code{1}. A value of e.g. \code{2} is used \emph{internally} sometimes.}
    \item{\code{min_sum}}{Minimum sum for the mixture coefficients. The maximum sum is
\code{1} (except when there is one coefficient only and \code{max_coef > 1}).
Default is \code{1 - 1e-8}, i.e. the sum must be very close to \code{1}.}
    \item{\code{nb_coef}}{Maximum number of non-zero mixture coefficients. Default is
\code{Inf} (no restriction).}
  }}
}
\value{
\code{W}, after convergence of the deconvolution
\eqn{PC \approx Q \cdot PC_\text{ref}} where \eqn{PC_\text{ref} = W^T PC}.
}
\description{
Alternating deconvolution with starting value.
}
\examples{
PC <- prcomp(iris[1:4])$x
PC_ref_init <- do.call("rbind", by(PC, iris$Species, colMeans))  # cheating
pc_plot(PC, PC_ref_init, color_var = iris$Species)

W <- pc_deconv_withstart(PC, PC_ref_init, m_exponent = 5)
PC_ref_conv <- pc_refs(PC, W)
pc_plot(PC, PC_ref_conv, color_var = iris$Species)

}
\seealso{
\link{pc_deconv}
}
