% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtures.R
\name{pc_mixtures}
\alias{pc_mixtures}
\title{Mixture coefficients}
\usage{
pc_mixtures(
  PC,
  PC_ref,
  min_coef = 1e-05,
  max_coef = 1,
  min_sum = 1 - 1e-08,
  nb_coef = Inf
)
}
\arguments{
\item{PC}{Matrix of principal components (N x K, where K is the number of PCs).}

\item{PC_ref}{Matrix of reference positions in the PCs (L x K, where L is the
number of references).}

\item{min_coef}{Minimum non-zero mixture coefficient. Default is \code{1e-5}.
This is used in the first iteration of the algorithm only, so it can happen
that some coefficients in the results are lower (but this should be rare).}

\item{max_coef}{Maximum mixture coefficient when only one coefficient is used.
Default is \code{1}. A value of e.g. \code{2} is used \emph{internally} sometimes.}

\item{min_sum}{Minimum sum for the mixture coefficients. The maximum sum is
\code{1} (except when there is one coefficient only and \code{max_coef > 1}).
Default is \code{1 - 1e-8}, i.e. the sum must be very close to \code{1}.}

\item{nb_coef}{Maximum number of non-zero mixture coefficients. Default is
\code{Inf} (no restriction).}
}
\value{
A matrix of mixture coefficients (N x L).
}
\description{
Mixture coefficients
}
\details{
A \code{future.apply} loop is used internally, which will use parallelism if you
registered a parallel backend with e.g. \code{future::plan}.
}
\examples{
PC <- prcomp(iris[1:4])$x
PC_ref <- do.call("rbind", by(PC, iris$Species, colMeans))
Q <- pc_mixtures(PC, PC_ref)

}
