% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rank_in_group}
\alias{rank_in_group}
\alias{pc_plot_mixtures}
\title{Title}
\usage{
rank_in_group(Q, group)

pc_plot_mixtures(
  Q,
  rank_in_group,
  colors = c("#64cb64", "#e7be28", "#b94a7b", "#6687d2", "#add042", "#cc4eb7", "#e6831f",
    "#8d5edf", "#5ccda0", "#613d9a", "#c79335", "#46d0e5", "#d14f29", "#60924e",
    "#a45441")
)
}
\arguments{
\item{group}{}

\item{colors}{}
}
\description{
Title

Title
}
\examples{
PC <- prcomp(iris[1:4])$x
PC_ref <- do.call("rbind", by(PC, iris$Species, colMeans))
Q <- pc_mixtures(PC, PC_ref)

rank_in_grp <- rank_in_group(Q, iris$Species)
pc_plot_mixtures(Q, rank_in_grp)

colors <- c("#5ccda0", "#613d9a", "#c79335")
colnames(Q) <- colors
pc_plot_mixtures(Q, rank_in_grp)
}
