% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdeconv-package.R
\docType{package}
\name{pcdeconv-package}
\alias{pcdeconv}
\alias{pcdeconv-package}
\title{pcdeconv: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
