% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pc_plot}
\alias{pc_plot}
\title{Title}
\usage{
pc_plot(
  PC,
  PC_ref = PC[0, ],
  nfacet = 9,
  color_var = I("black"),
  color_ref = "red",
  legend_ratio = 0.2,
  ...
)
}
\arguments{
\item{...}{Further parameters to pass to \link[cowplot:plot_grid]{cowplot::plot_grid}}
}
\description{
Title
}
