% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pc_plot_mixtures}
\alias{pc_plot_mixtures}
\title{Title}
\usage{
pc_plot_mixtures(
  Q,
  rank_in_group,
  colors = c("#64cb64", "#e7be28", "#b94a7b", "#6687d2", "#add042", "#cc4eb7", "#e6831f",
    "#8d5edf", "#5ccda0", "#613d9a", "#c79335", "#46d0e5", "#d14f29", "#60924e",
    "#a45441")
)
}
\arguments{
\item{colors}{}
}
\description{
Title
}
