% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mixtures.R
\docType{data}
\name{pc_plot_mixtures}
\alias{pc_plot_mixtures}
\alias{rank_in_group}
\alias{COLORS_DECONV}
\title{Plot mixtures}
\format{
COLORS_DECONV: a vector of 30 colors.
}
\usage{
pc_plot_mixtures(Q, rank_in_group, colors = COLORS_DECONV)

rank_in_group(Q, group)

COLORS_DECONV
}
\arguments{
\item{Q}{Matrix of mixture coefficients (L x K), typically the output
of \code{\link[=pc_mixtures]{pc_mixtures()}}.}

\item{rank_in_group}{The output of \code{\link[=rank_in_group]{rank_in_group()}}.}

\item{colors}{Colors to use for references. You can assign these colors to
colnames of \code{Q} to assign specific colors to specific references.}

\item{group}{Vector of labels to use for facetting in \code{\link[=pc_plot_mixtures]{pc_plot_mixtures()}}.}
}
\value{
A ggplot object.
}
\description{
Plot mixtures
}
\examples{
PC <- prcomp(iris[1:4])$x
PC_ref <- do.call("rbind", by(PC, iris$Species, colMeans))
Q <- pc_mixtures(PC, PC_ref)

rank_in_grp <- rank_in_group(Q, iris$Species)
pc_plot_mixtures(Q, rank_in_grp)

colors <- c("#5ccda0", "#613d9a", "#c79335")
colnames(Q) <- colors
pc_plot_mixtures(Q, rank_in_grp)
pc_plot_mixtures(Q, rank_in_grp, colors = rev(colors))
pc_plot_mixtures(Q, rank_in_grp) +
  ggplot2::facet_wrap(~ .GRP, ncol = 1, scales = "free_x")
}
\keyword{internal}
