% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rank_in_group}
\alias{rank_in_group}
\alias{pc_plot_mixtures}
\title{Plot mixtures}
\usage{
rank_in_group(Q, group)

pc_plot_mixtures(
  Q,
  rank_in_group,
  colors = c("#64cb64", "#e7be28", "#d14f29", "#6687d2", "#add042", "#cc4eb7", "#e6831f",
    "#c5b0d5", "#5ccda0", "#613d9a", "#c79335", "#46d0e5", "#b94a7b", "#60924e",
    "#a45441", "#f7b6d2", "#d62728", "#8c564b", "#e377c2", "#bcbd22", "#17becf",
    "#aec7e8", "#ffbb78", "#98df8a", "#ff9896", "#8d5edf", "#c49c94", "#1f77b4",
    "#7f7f7f", "#000000")
)
}
\arguments{
\item{Q}{Matrix of mixture coefficients (L x K), typically the output
of \code{\link[=pc_mixtures]{pc_mixtures()}}.}

\item{group}{Vector of labels to use for facetting in \code{\link[=pc_plot_mixtures]{pc_plot_mixtures()}}.}

\item{rank_in_group}{The output of \code{\link[=rank_in_group]{rank_in_group()}}.}

\item{colors}{}
}
\value{
A ggplot object.
}
\description{
Plot mixtures
}
\examples{
PC <- prcomp(iris[1:4])$x
PC_ref <- do.call("rbind", by(PC, iris$Species, colMeans))
Q <- pc_mixtures(PC, PC_ref)

rank_in_grp <- rank_in_group(Q, iris$Species)
pc_plot_mixtures(Q, rank_in_grp)

colors <- c("#5ccda0", "#613d9a", "#c79335")
colnames(Q) <- colors
pc_plot_mixtures(Q, rank_in_grp)
pc_plot_mixtures(Q, rank_in_grp, colors = rev(colors))
pc_plot_mixtures(Q, rank_in_grp) + facet_wrap(~ .GRP, ncol = 1, scales = "free_x")
}
