% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs.R
\name{pc_weights_refs}
\alias{pc_weights_refs}
\alias{pc_refs}
\title{Weights to compute references}
\usage{
pc_weights_refs(Q, m_exponent, thr_coef = 0.6)

pc_refs(PC, W)
}
\arguments{
\item{Q}{Matrix of mixture coefficients (L x K), typically the output
of \code{\link[=pc_mixtures]{pc_mixtures()}}.}

\item{m_exponent}{Exponent to be applied to the coefficients of \code{Q}.
A value between 3 and 30 is usually appropriate to priotize larger Q values.}

\item{thr_coef}{Threshold under which coefficients of \code{Q} are not used.
Default is 0.6.}

\item{PC}{Matrix of PC scores, N x K.}

\item{W}{The output of \code{\link[=pc_weights_refs]{pc_weights_refs()}}.}
}
\value{
Sparse matrix \code{W} of weights (N x L) to be used to compute reference
positions in PCs as \code{PC_ref <- pc_refs(PC, W)}.
}
\description{
Weights to compute references
}
\examples{
PC <- prcomp(iris[1:4])$x
PC_ref <- do.call("rbind", by(PC, iris$Species, colMeans))
pc_plot(PC, PC_ref, color_var = iris$Species)

Q <- pc_mixtures(PC, PC_ref)
W <- pc_weights_refs(Q, m_exponent = 10)
PC_ref2 <- pc_refs(PC, W)
pc_plot(PC, PC_ref2, color_var = iris$Species)

}
